move ProfileImage to components/display
