+++ src/components/legacy/ListItem/ListItemWithImage.js
@@ -25,7 +25,7 @@ import Icon from 'components/legacy/Icon';
 import t from 'translations/translate';
 
 import { BaseText, MediumText } from 'components/legacy/Typography';
-import ProfileImage from 'components/ProfileImage';
+import ProfileImage from 'components/display/ProfileImage';
 import Button from 'components/legacy/Button';
 import Image from 'components/core/Image';
 import { Shadow } from 'components/Shadow';
